refactor: extract alc-tenko crate + move middleware to alc-auth (Phase 6)
- Move middleware functions (require_jwt, require_tenant) to alc-auth
- Extract 8 tenko route handlers to alc-tenko crate:
  tenko_sessions, tenko_records, tenko_schedules, tenko_call,
  tenko_webhooks, health_baselines, equipment_failures, daily_health
- Main crate re-exports via pub use for backward compatibility
