+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "crates/alc-auth", "crates/alc-core", "crates/alc-csv-parser", "crates/alc-compare", "crates/alc-pdf", "crates/alc-storage"]
+members = [".", "crates/alc-auth", "crates/alc-core", "crates/alc-csv-parser", "crates/alc-compare", "crates/alc-pdf", "crates/alc-storage", "crates/alc-tenko"]
 resolver = "2"
 
 [package]
@@ -13,6 +13,7 @@ unexpected_cfgs = { level = "allow", check-cfg = ['cfg(coverage)'] }
 [dependencies]
 alc-auth = { path = "crates/alc-auth" }
 alc-core = { path = "crates/alc-core" }
+alc-tenko = { path = "crates/alc-tenko" }
 alc-storage = { path = "crates/alc-storage" }
 alc-csv-parser = { path = "crates/alc-csv-parser" }
 alc-compare = { path = "crates/alc-compare" }
