+++ crates/alc-auth/Cargo.toml
@@ -8,6 +8,7 @@ unexpected_cfgs = { level = "allow", check-cfg = ['cfg(coverage)'] }
 
 [dependencies]
 alc-core = { path = "../alc-core" }
+axum = "0.8"
 base64 = "0.22"
 chrono = { version = "0.4", features = ["serde"] }
 hmac = "0.12"
