+++ crates/alc-auth/src/lib.rs
@@ -5,3 +5,4 @@ mod test_macros;
 pub mod google;
 pub mod jwt;
 pub mod lineworks;
+pub mod middleware;
