+++ crates/alc-tenko/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "alc-tenko"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+alc-core = { path = "../alc-core" }
+alc-auth = { path = "../alc-auth" }
+axum = "0.8"
+chrono = { version = "0.4", features = ["serde"] }
+csv = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
+tokio = "1"
+tracing = "0.1"
+uuid = { version = "1", features = ["v4", "serde"] }
