+++ crates/alc-tenko/src/tenko_records.rs
@@ -8,9 +8,9 @@ use axum::{
 };
 use uuid::Uuid;
 
-use crate::db::models::{TenkoRecord, TenkoRecordFilter, TenkoRecordsResponse};
-use crate::middleware::auth::TenantId;
-use crate::AppState;
+use alc_auth::middleware::TenantId;
+use alc_core::models::{TenkoRecord, TenkoRecordFilter, TenkoRecordsResponse};
+use alc_core::AppState;
 
 /// テナント対応ルート (JWT or X-Tenant-ID)
 pub fn tenant_router() -> Router<AppState> {
