+++ src/routes/mod.rs
@@ -5,7 +5,7 @@ pub mod car_inspections;
 pub mod carins_files;
 pub mod carrying_items;
 pub mod communication_items;
-pub mod daily_health;
+pub use alc_tenko::daily_health;
 pub mod devices;
 pub mod driver_info;
 pub mod dtako_csv_proxy;
@@ -20,19 +20,19 @@ pub mod dtako_upload;
 pub mod dtako_vehicles;
 pub mod dtako_work_times;
 pub mod employees;
-pub mod equipment_failures;
+pub use alc_tenko::equipment_failures;
 pub mod guidance_records;
 pub mod health;
-pub mod health_baselines;
+pub use alc_tenko::health_baselines;
 pub mod measurements;
 pub mod nfc_tags;
 pub mod sso_admin;
 pub mod tenant_users;
-pub mod tenko_call;
-pub mod tenko_records;
-pub mod tenko_schedules;
-pub mod tenko_sessions;
-pub mod tenko_webhooks;
+pub use alc_tenko::tenko_call;
+pub use alc_tenko::tenko_records;
+pub use alc_tenko::tenko_schedules;
+pub use alc_tenko::tenko_sessions;
+pub use alc_tenko::tenko_webhooks;
 pub mod timecard;
 pub mod upload;
